requirements.txt
Flask==1.0.2
google-auth==1.11.2
requests==2.22.0